core: add core of transform layer with steps
This is for Flux, which probably is not as important as other
use cases since it understands a jobspec, but it demonstrates
the basic staging, writing, and submit of a script. I will
add other steps I need now, and then likely other transformers
for things that are not flux.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>