spec: add first draft for jobspec nextgen
Problem: we need to start working on the
nextgen jobspec!
Solution: create a spec.md that does that.

I will next start prototyping this into the
library. My first goal will be taking jobspecs
and running work with flux (to build, run, etc)
and then we can map this into flux miniclusters
with the rainbow scheduler.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>