wip: adding updated spec with groups
Problem: we did not have a way to describe batch vs submit
Solution: a batch is a group of jobs or tasks. If we define
an abstraction of groups, it is easy to delineate the two,
and the user is free to still define a-la-carte tasks, tasks
within groups, or groups within tasks.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>