feat: support for flux workload
Problem: a workload can be any number of submit/batch
Solution: create a workload abstraction that can handle
the nesting. We can write flux batch scripts as jobspecs
to submit with flux job submit, and we do this by doing
flux submit with --dry-run and then adding a faux command
to a batch-script in files that does not exist. Then we load
the jobspec, add the faux script (with the subtasks) and then
flux job submit <jobspec>. Next I need to work on the group
in group case (a flux batch that submits another flux batch)

Signed-off-by: vsoch <vsoch@users.noreply.github.com>