refactor: represent requires in resources
Problem: a compatibility requirement needs to clearly be
matched to resources, at least the subset that goes into
a match in the graph.
Solution: refactor the jobspec so top level resources can
have slots that have nested "requires" at the level they
are needed. I also found a way to remove the "slot" idea
and just replace it with the idea of asking for replicas of
a particular place in the resource request. Note that
the implementation is not updated yet, just the spec, and
I am opening a PR so I can stare at it.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>