refactor: represent requires in resources (#14)
* refactor: represent requires in resources

Problem: a compatibility requirement needs to clearly be
matched to resources, at least the subset that goes into
a match in the graph.
Solution: refactor the jobspec so top level resources can
have slots that have nested "requires" at the level they
are needed. I also found a way to remove the "slot" idea
and just replace it with the idea of asking for replicas of
a particular place in the resource request. Note that
the implementation is not updated yet, just the spec, and
I am opening a PR so I can stare at it.

* schema: refactor schema

Problem: requires needs to be grouped with schedule-able
groups.
Solution: simplify the jobspec to allow that.

* requires should be array of objects
* update example with requires array
* update requires in spec
* do not require enum of just node for resources
* allow attributes on the top level of jobspec

These are for cluster state that is desired. We likely
will want to move these into schedulable units.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>